Modified return value to int

+++ src/main/java/abramyan/begin/Begin5.java
@@ -36,7 +36,7 @@ public class Begin5 {
         System.out.println(calculateSurfaceArea(a));
     }
 
-    private static double calculateVolume(int a) {
+    private static int calculateVolume(int a) {
         return a * a * a;
     }
